+++ .env
@@ -1,0 +1,2 @@
+REACT_APP_API_URL=https://opendata.resas-portal.go.jp/api/v1
+REACT_APP_API_KEY=dTculG00CntfRMFfEnc3feAkbI0Y8TKiojhI2dUY
