+++ src/components/Top.css
@@ -1,0 +1,44 @@
+.prefecture_title {
+  font-size: 30px;
+  margin: 10px 0;
+  @media screen and (max-width: 425px) {
+    font-size: 20px;
+    margin: 5px 0;
+  }
+}
+
+.prefecture_list {
+  list-style: none;
+  display: flex;
+  justify-content: flex-start;
+  flex-wrap: wrap;
+  padding: 0;
+}
+
+.prefecture_list_item {
+  margin: 3px 10px;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+  padding: 5px 10px;
+  @media screen and (max-width: 425px) {
+    margin: 3px 5px;
+    padding: 5px 5px;
+  }
+}
+
+.prefecture_list_item_label {
+  font-size: 18px;
+  @media screen and (max-width: 425px) {
+    font-size: 10px;
+  }
+}
+
+.prefecture_list_item:hover {
+  background-color: #eee;
+}
+
+.prefecture_list_item:hover,
+.prefecture_list_item_label:hover,
+.prefecture_list_item_checkbox:hover {
+  cursor: pointer;
+}
