+++ src/globalStates/recoilKeys.ts
@@ -1,0 +1,7 @@
+export enum AtomKeys {
+  PREFECTURES_STATE = 'prefecturesState',
+}
+
+export enum SelectorKeys {
+  CHAR_COUNT_STATE = 'charCountState',
+}
