+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export type PrefectureType = {
+  prefCode: number;
+  prefName: string;
+};
